by gacp

+++ Module-07/ex02/main.cpp
@@ -6,7 +6,7 @@
 /*   By: youmoukh <youmoukh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/05 16:42:01 by youmoukh          #+#    #+#             */
-/*   Updated: 2024/10/12 14:24:43 by youmoukh         ###   ########.fr       */
+/*   Updated: 2024/10/12 14:25:00 by youmoukh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ int main(int you, char **nes)
 	B = A;
 	std::cout << B.getSize() << std::endl;
 	
-	for (int i = 0; i < B.getSize() ; i++)
+	for (size_t i = 0; i < B.getSize() ; i++)
 	{
 		std::cout <<  B[i] << std::endl;
 	}
